detect: move rule loading into loader files
