+++ src/detect.h
@@ -1177,14 +1177,6 @@ typedef struct DetectEngineMasterCtx_ {
     int keyword_id;
 } DetectEngineMasterCtx;
 
-/** \brief Signature loader statistics */
-typedef struct SigFileLoaderStat_ {
-    int bad_files;
-    int total_files;
-    int good_sigs_total;
-    int bad_sigs_total;
-} SigFileLoaderStat;
-
 /** Remember to add the options in SignatureIsIPOnly() at detect.c otherwise it wont be part of a signature group */
 
 /* detection api */
